# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for the gdsi API",
        "title": "Gdsi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/players": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get players",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/players.PlayerModel"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/failure.ValidationFailure"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    }
                }
            }
        },
        "/v1/players/{playerId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get player by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player id",
                        "name": "playerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/matches.MatchModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/failure.ValidationFailure"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player id",
                        "name": "playerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/players.UpdatePlayerRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/players.PlayerModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/failure.ValidationFailure"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "failure.Failure": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "failure.InvalidField": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "location": {
                    "description": "location specifies where the field is conmming from (path, query, body)",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "failure.ValidationFailure": {
            "type": "object",
            "properties": {
                "invalid_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/failure.InvalidField"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "matches.CourtModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "matches.LeagueModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "matches.MatchModel": {
            "type": "object",
            "properties": {
                "court": {
                    "$ref": "#/definitions/matches.CourtModel"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "league": {
                    "$ref": "#/definitions/matches.LeagueModel"
                },
                "player_one": {
                    "$ref": "#/definitions/matches.PlayerModel"
                },
                "player_two": {
                    "$ref": "#/definitions/matches.PlayerModel"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "score": {
                    "type": "string"
                },
                "season": {
                    "$ref": "#/definitions/matches.SeasonModel"
                },
                "winner": {
                    "$ref": "#/definitions/matches.PlayerModel"
                }
            }
        },
        "matches.PlayerModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "matches.SeasonModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "players.AccountModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "players.CurrentLeagueModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "players.PlayerModel": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/players.AccountModel"
                },
                "created_at": {
                    "type": "string"
                },
                "current_league": {
                    "$ref": "#/definitions/players.CurrentLeagueModel"
                },
                "handedness": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "matches_expected": {
                    "type": "integer"
                },
                "matches_played": {
                    "type": "integer"
                },
                "matches_scheduled": {
                    "type": "integer"
                },
                "matches_won": {
                    "type": "integer"
                },
                "racket": {
                    "type": "string"
                },
                "seasons_played": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "players.UpdatePlayerRequestModel": {
            "type": "object",
            "properties": {
                "handedness": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "racket": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the Bearer token in the format: Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}